+++ tests/e2e/client/aptosRequest.test.ts
@@ -1,28 +1,31 @@
 import {
-  aptosRequest,
-  AptosApiError,
-  AptosConfig,
-  Network,
-  NetworkToNodeAPI,
-  Aptos,
   Account,
-  U64,
+  AptosApiError,
+  AptosApiType,
+  aptosRequest,
+  generateSignedTransaction,
   GraphqlQuery,
   NetworkToIndexerAPI,
-  generateSignedTransaction,
+  NetworkToNodeAPI,
+  U64,
 } from "../../../src";
 import { VERSION } from "../../../src/version";
 import { longTestTimeout } from "../../unit/helper";
+import { getAptosClient } from "../helper";
 import { singleSignerScriptBytecode } from "../transaction/helper";
 
-const config = new AptosConfig({ network: Network.LOCAL });
+const { aptos, config } = getAptosClient();
+
+const fullnodeUrl = config.fullnode ?? NetworkToNodeAPI[config.network];
+const indexerUrl = config.indexer ?? NetworkToIndexerAPI[config.network];
+
+const dummyKey = "aptoslabs_secret1";
 
 describe("aptos request", () => {
   describe("headers", () => {
     test(
       "call should include all expected headers",
       async () => {
-        const aptos = new Aptos(config);
         const sender = Account.generate();
         const receiverAccounts = Account.generate();
         await aptos.fundAccount({ accountAddress: sender.accountAddress, amount: 100_000_000 });
@@ -41,7 +44,7 @@ describe("aptos request", () => {
         try {
           const response = await aptosRequest(
             {
-              url: `${NetworkToNodeAPI[config.network]}`,
+              url: fullnodeUrl,
               method: "POST",
               path: "transactions",
               body: signedTransaction,
@@ -50,57 +53,11 @@ describe("aptos request", () => {
               overrides: { HEADERS: { my: "header" } },
             },
             config,
+            AptosApiType.FULLNODE,
           );
           expect(response.config.headers).toHaveProperty("x-aptos-client", `aptos-typescript-sdk/${VERSION}`);
           expect(response.config.headers).toHaveProperty("my", "header");
           expect(response.config.headers).toHaveProperty("content-type", "application/x.aptos.signed_transaction+bcs");
-        } catch (error: any) {
-          // should not get here
-          expect(true).toBe(false);
-        }
-      },
-      longTestTimeout,
-    );
-  });
-
-  describe("token", () => {
-    test(
-      "should not set auth_token for full node requests",
-      async () => {
-        try {
-          const response = await aptosRequest(
-            {
-              url: `${NetworkToNodeAPI[config.network]}`,
-              method: "GET",
-              path: "accounts/0x1",
-              overrides: { AUTH_TOKEN: "my-token" },
-              originMethod: "test when token is set",
-            },
-            config,
-          );
-          expect(response.config.headers).not.toHaveProperty("authorization", "Bearer my-token");
-        } catch (error: any) {
-          // should not get here
-          expect(true).toBe(false);
-        }
-      },
-      longTestTimeout,
-    );
-
-    test(
-      "when token is not set",
-      async () => {
-        try {
-          const response = await aptosRequest(
-            {
-              url: `${NetworkToNodeAPI[config.network]}`,
-              method: "GET",
-              path: "accounts/0x1",
-              originMethod: "test when token is not set",
-            },
-            config,
-          );
-          expect(response.config.headers).not.toHaveProperty("authorization", "Bearer my-token");
         } catch (error: any) {
           // should not get here
           expect(true).toBe(false);
@@ -117,15 +74,16 @@ describe("aptos request", () => {
         try {
           const response = await aptosRequest(
             {
-              url: `${NetworkToNodeAPI[config.network]}`,
+              url: fullnodeUrl,
               method: "GET",
               path: "accounts/0x1",
-              overrides: { API_KEY: "my-api-key" },
+              overrides: { API_KEY: dummyKey },
               originMethod: "test when token is set",
             },
             config,
+            AptosApiType.FULLNODE,
           );
-          expect(response.config.headers).toHaveProperty("authorization", "Bearer my-api-key");
+          expect(response.config.headers).toHaveProperty("authorization", `Bearer ${dummyKey}`);
         } catch (error: any) {
           // should not get here
           expect(true).toBe(false);
@@ -143,12 +101,13 @@ describe("aptos request", () => {
           try {
             const response = await aptosRequest(
               {
-                url: `${NetworkToNodeAPI[config.network]}`,
+                url: fullnodeUrl,
                 method: "GET",
                 path: "accounts/0x1",
                 originMethod: "test fullnode 200 status",
               },
               config,
+              AptosApiType.FULLNODE,
             );
             expect(response).toHaveProperty("data", {
               sequence_number: "0",
@@ -170,16 +129,17 @@ describe("aptos request", () => {
           try {
             await aptosRequest(
               {
-                url: `${NetworkToNodeAPI[config.network]}`,
+                url: fullnodeUrl,
                 method: "GET",
                 path: "transactions/by_hash/0x123",
                 originMethod: "test 400 status",
               },
               config,
+              AptosApiType.FULLNODE,
             );
           } catch (error: any) {
             expect(error).toBeInstanceOf(AptosApiError);
-            expect(error.url).toBe(`${NetworkToNodeAPI[config.network]}/transactions/by_hash/0x123`);
+            expect(error.url).toBe(`${fullnodeUrl}/transactions/by_hash/0x123`);
             expect(error.status).toBe(400);
             expect(error.statusText).toBe("Bad Request");
             expect(error.data).toEqual({
@@ -190,7 +150,7 @@ describe("aptos request", () => {
               vm_error_code: null,
             });
             expect(error.request).toEqual({
-              url: `${NetworkToNodeAPI[config.network]}`,
+              url: `${fullnodeUrl}`,
               method: "GET",
               originMethod: "test 400 status",
               path: "transactions/by_hash/0x123",
@@ -205,19 +165,18 @@ describe("aptos request", () => {
           try {
             await aptosRequest(
               {
-                url: `${NetworkToNodeAPI[config.network]}`,
+                url: `${fullnodeUrl}`,
                 method: "GET",
                 path: "transactions/by_hash/0x23851af73879128b541bafad4b49d0b6f1ac0d49ed2400632d247135fbca7bea",
                 originMethod: "test 404 status",
               },
               config,
+              AptosApiType.FULLNODE,
             );
           } catch (error: any) {
             expect(error).toBeInstanceOf(AptosApiError);
             expect(error.url).toBe(
-              `${
-                NetworkToNodeAPI[config.network]
-              }/transactions/by_hash/0x23851af73879128b541bafad4b49d0b6f1ac0d49ed2400632d247135fbca7bea`,
+              `${fullnodeUrl}/transactions/by_hash/0x23851af73879128b541bafad4b49d0b6f1ac0d49ed2400632d247135fbca7bea`,
             );
             expect(error.status).toBe(404);
             expect(error.statusText).toBe("Not Found");
@@ -228,7 +187,7 @@ describe("aptos request", () => {
               vm_error_code: null,
             });
             expect(error.request).toEqual({
-              url: `${NetworkToNodeAPI[config.network]}`,
+              url: `${fullnodeUrl}`,
               method: "GET",
               originMethod: "test 404 status",
               path: "transactions/by_hash/0x23851af73879128b541bafad4b49d0b6f1ac0d49ed2400632d247135fbca7bea",
@@ -244,7 +203,7 @@ describe("aptos request", () => {
           try {
             await aptosRequest(
               {
-                url: `${NetworkToNodeAPI[config.network]}`,
+                url: `${fullnodeUrl}`,
                 method: "POST",
                 path: "transactions",
                 body: new Uint8Array([1, 2, 3]),
@@ -252,10 +211,11 @@ describe("aptos request", () => {
                 contentType: "application/x.aptos.signed_transaction+bcs",
               },
               config,
+              AptosApiType.FULLNODE,
             );
           } catch (error: any) {
             expect(error).toBeInstanceOf(AptosApiError);
-            expect(error.url).toBe(`${NetworkToNodeAPI[config.network]}/transactions`);
+            expect(error.url).toBe(`${fullnodeUrl}/transactions`);
             expect(error.status).toBe(400);
             expect(error.statusText).toBe("Bad Request");
             expect(error.data).toEqual({
@@ -264,7 +224,7 @@ describe("aptos request", () => {
               vm_error_code: null,
             });
             expect(error.request).toEqual({
-              url: `${NetworkToNodeAPI[config.network]}`,
+              url: `${fullnodeUrl}`,
               method: "POST",
               originMethod: "test transaction submission error",
               path: "transactions",
@@ -293,12 +253,13 @@ describe("aptos request", () => {
             };
             const response = await aptosRequest(
               {
-                url: `${NetworkToIndexerAPI[config.network]}`,
+                url: `${indexerUrl}`,
                 method: "POST",
                 body: query,
                 originMethod: "test indexer 200 status",
               },
               config,
+              AptosApiType.INDEXER,
             );
             expect(response).toHaveProperty("data", {
               ledger_infos: [
@@ -329,16 +290,17 @@ describe("aptos request", () => {
             };
             await aptosRequest(
               {
-                url: `${NetworkToIndexerAPI[config.network]}`,
+                url: `${indexerUrl}`,
                 method: "POST",
                 body: query,
                 originMethod: "test indexer 400 status",
               },
               config,
+              AptosApiType.INDEXER,
             );
           } catch (error: any) {
             expect(error).toBeInstanceOf(AptosApiError);
-            expect(error.url).toBe(`${NetworkToIndexerAPI[config.network]}`);
+            expect(error.url).toBe(`${indexerUrl}`);
             expect(error.status).toBe(200);
             expect(error.statusText).toBe("OK");
             expect(error.data).toHaveProperty("errors");
@@ -349,7 +311,7 @@ describe("aptos request", () => {
               },
             ]);
             expect(error.request).toEqual({
-              url: `${NetworkToIndexerAPI[config.network]}`,
+              url: `${indexerUrl}`,
               method: "POST",
               originMethod: "test indexer 400 status",
               body: {